Setup clovery service serializer and reorganize folders
