+++ lib/serializers/carbonfootprint_serializer.rb
@@ -1,0 +1,6 @@
+class CarbonfootprintSerializer
+  inlude FastJsonapi::ObjectSerializer
+  attributes :equivalent_carbon_in_kg,
+             :currency,
+             :total
+end
