+++ lib/services/cloverly_service.rb
