+++ spec/spec_helper.rb
@@ -3,5 +3,5 @@ require 'rspec'
 
 SimpleCov.start
 
-require File.expand_path '../lib/cloverly_service.rb', __dir__
+require File.expand_path '../lib/services/cloverly_service.rb', __dir__
 require File.expand_path '../lib/poros/carbonfootprint.rb', __dir__
